some code restructuring + comments; no logic changes yet
