+++ src/domain/space/admin/SpaceCommunity/AdminSpaceCommunityPage.tsx
@@ -9,8 +9,8 @@ import { useSpace } from '../../context/useSpace';
 import PageContent from '@/core/ui/content/PageContent';
 import PageContentBlock from '@/core/ui/content/PageContentBlock';
 import PageContentColumn from '@/core/ui/content/PageContentColumn';
-import CommunityUsers from '@/domain/community/community/CommunityAdmin/CommunityUsers';
-import useCommunityAdmin from '@/domain/community/community/CommunityAdmin/useCommunityAdmin';
+import CommunityUsers from '@/domain/space/admin/SpaceCommunity/CommunityUsers';
+import useCommunityAdmin from '@/domain/space/admin/SpaceCommunity/useCommunityAdmin';
 import CommunityOrganizations from '@/domain/community/community/CommunityAdmin/CommunityOrganizations';
 import CommunityApplications from '@/domain/community/community/CommunityAdmin/CommunityApplications';
 import PageContentBlockSeamless from '@/core/ui/content/PageContentBlockSeamless';
