+++ src/domain/space/admin/SpaceCommunity/AdminSubspaceCommunityPage.tsx
@@ -6,8 +6,8 @@ import PageContentBlock from '@/core/ui/content/PageContentBlock';
 import PageContentColumn from '@/core/ui/content/PageContentColumn';
 import CommunityApplications from '@/domain/community/community/CommunityAdmin/CommunityApplications';
 import CommunityOrganizations from '@/domain/community/community/CommunityAdmin/CommunityOrganizations';
-import CommunityUsers from '@/domain/community/community/CommunityAdmin/CommunityUsers';
-import useCommunityAdmin from '@/domain/community/community/CommunityAdmin/useCommunityAdmin';
+import CommunityUsers from '@/domain/space/admin/SpaceCommunity/CommunityUsers';
+import useCommunityAdmin from '@/domain/space/admin/SpaceCommunity/useCommunityAdmin';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import { useSubSpace } from '../../hooks/useSubSpace';
