+++ src/domain/space/admin/SpaceCommunity/SpaceEntitlements.graphql
